+++ internal/amf/ngap/handle_pdu_session_resource_setup_response.go
@@ -82,11 +82,11 @@ func HandlePDUSessionResourceSetupResponse(ctx ctxt.Context, ran *context.AmfRan
 			ranUe.Log.Debug("Send PDUSessionResourceSetupResponseTransfer to SMF")
 
 			for _, item := range pDUSessionResourceSetupResponseList.List {
-				pduSessionID := int32(item.PDUSessionID.Value)
+				pduSessionID := uint8(item.PDUSessionID.Value)
 				transfer := item.PDUSessionResourceSetupResponseTransfer
 				smContext, ok := amfUe.SmContextFindByPDUSessionID(pduSessionID)
 				if !ok {
-					ranUe.Log.Error("SmContext not found", zap.Int32("PduSessionID", pduSessionID))
+					ranUe.Log.Error("SmContext not found", zap.Uint8("PduSessionID", pduSessionID))
 					continue
 				}
 				response, err := consumer.SendUpdateSmContextN2Info(ctx, amfUe, smContext,
@@ -106,11 +106,11 @@ func HandlePDUSessionResourceSetupResponse(ctx ctxt.Context, ran *context.AmfRan
 			ranUe.Log.Debug("Send PDUSessionResourceSetupUnsuccessfulTransfer to SMF")
 
 			for _, item := range pDUSessionResourceFailedToSetupList.List {
-				pduSessionID := int32(item.PDUSessionID.Value)
+				pduSessionID := uint8(item.PDUSessionID.Value)
 				transfer := item.PDUSessionResourceSetupUnsuccessfulTransfer
 				smContext, ok := amfUe.SmContextFindByPDUSessionID(pduSessionID)
 				if !ok {
-					ranUe.Log.Error("SmContext not found", zap.Int32("PduSessionID", pduSessionID))
+					ranUe.Log.Error("SmContext not found", zap.Uint8("PduSessionID", pduSessionID))
 				}
 				_, err := consumer.SendUpdateSmContextN2Info(ctx, amfUe, smContext, models.N2SmInfoTypePduResSetupFail, transfer)
 				if err != nil {
